Create and "EMS" module for order execution/routing actor(s)
